refactor(audio-bridge): msg -> options
