+++ jarust_plugins/examples/audio_bridge_allowed.rs
@@ -1,9 +1,9 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::AllowAction;
-use jarust_plugins::audio_bridge::messages::AllowedMsg;
-use jarust_plugins::audio_bridge::messages::CreateRoomMsg;
+use jarust_plugins::audio_bridge::msg_opitons::AllowAction;
+use jarust_plugins::audio_bridge::msg_opitons::AllowedOptions;
+use jarust_plugins::audio_bridge::msg_opitons::CreateRoomOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
 
@@ -24,7 +24,7 @@ async fn main() -> anyhow::Result<()> {
 
     let create_room_rsp = handle
         .create_room_with_config(
-            CreateRoomMsg {
+            CreateRoomOptions {
                 secret: Some("superdupersecret".to_string()),
                 ..Default::default()
             },
@@ -40,7 +40,7 @@ async fn main() -> anyhow::Result<()> {
     let allowed_rsp = handle
         .allowed(
             create_room_rsp.room,
-            AllowedMsg {
+            AllowedOptions {
                 action: AllowAction::Add,
                 allowed: vec![],
                 secret: Some("superdupersecret".to_string()),
