+++ jarust_plugins/examples/audio_bridge_configure.rs
@@ -1,7 +1,7 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::ConfigureMsg;
+use jarust_plugins::audio_bridge::msg_opitons::ConfigureOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
 
@@ -29,7 +29,7 @@ async fn main() -> anyhow::Result<()> {
 
     handle
         .configure(
-            ConfigureMsg {
+            ConfigureOptions {
                 muted: Some(true),
                 display: Some("A configured display".to_string()),
                 ..Default::default()
