+++ jarust_plugins/examples/audio_bridge_create_room.rs
@@ -1,7 +1,7 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::CreateRoomMsg;
+use jarust_plugins::audio_bridge::msg_opitons::CreateRoomOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
 
@@ -21,7 +21,7 @@ async fn main() -> anyhow::Result<()> {
 
     let create_room_rsp = handle
         .create_room_with_config(
-            CreateRoomMsg {
+            CreateRoomOptions {
                 secret: Some("superdupersecret".to_string()),
                 ..Default::default()
             },
