+++ jarust_plugins/examples/audio_bridge_edit.rs
@@ -2,8 +2,8 @@ use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::common::Identifier;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::CreateRoomMsg;
-use jarust_plugins::audio_bridge::messages::EditRoomMsg;
+use jarust_plugins::audio_bridge::msg_opitons::CreateRoomOptions;
+use jarust_plugins::audio_bridge::msg_opitons::EditRoomOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
 
@@ -23,7 +23,7 @@ async fn main() -> anyhow::Result<()> {
 
     let _ = handle
         .create_room_with_config(
-            CreateRoomMsg {
+            CreateRoomOptions {
                 room: Some(Identifier::Uint(4321)),
                 description: Some("A nice description".to_string()),
                 secret: Some("superdupersecret".to_string()),
@@ -36,7 +36,7 @@ async fn main() -> anyhow::Result<()> {
     let edit_room_rsp = handle
         .edit_room(
             Identifier::Uint(4321),
-            EditRoomMsg {
+            EditRoomOptions {
                 new_description: Some("A nicer description".to_string()),
                 secret: Some("superdupersecret".to_string()),
                 ..Default::default()
