+++ jarust_plugins/examples/audio_bridge_exists.rs
@@ -2,8 +2,8 @@ use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::common::Identifier;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::CreateRoomMsg;
-use jarust_plugins::audio_bridge::messages::DestroyRoomMsg;
+use jarust_plugins::audio_bridge::msg_opitons::CreateRoomOptions;
+use jarust_plugins::audio_bridge::msg_opitons::DestroyRoomMsg;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
 
@@ -27,7 +27,7 @@ async fn main() -> anyhow::Result<()> {
     if !exist_rsp {
         let _ = handle
             .create_room_with_config(
-                CreateRoomMsg {
+                CreateRoomOptions {
                     room: Some(Identifier::Uint(4321)),
                     secret: Some("superdupersecret".to_string()),
                     ..Default::default()
