+++ jarust_plugins/examples/audio_bridge_mute_unmute.rs
@@ -1,7 +1,7 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::MuteOptions;
+use jarust_plugins::audio_bridge::msg_opitons::MuteOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
