+++ jarust_plugins/examples/audio_bridge_mute_unmute_room.rs
@@ -1,7 +1,7 @@
 use jarust::jaconfig::JaConfig;
 use jarust::jaconfig::TransportType;
 use jarust_plugins::audio_bridge::jahandle_ext::AudioBridge;
-use jarust_plugins::audio_bridge::messages::MuteRoomOptions;
+use jarust_plugins::audio_bridge::msg_opitons::MuteRoomOptions;
 use std::path::Path;
 use tracing_subscriber::EnvFilter;
