+++ jarust_plugins/src/audio_bridge/mod.rs
@@ -2,5 +2,5 @@ pub mod common;
 pub mod events;
 pub mod handle;
 pub mod jahandle_ext;
-pub mod messages;
+pub mod msg_opitons;
 pub mod responses;
